+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added <podcast:medium> support.
 - Added <podcast:locked> support.
+- Added <podcast:funding> support.
 
 ## 0.7.0 - 2023-05-23